Add toString() formatting of CqlDuration
Add formatting of CqlDuration::toString(), i.e. 1y2d3h. This
implementation is ported from Scylla source code.

+++ scylla-cdc-base/src/main/java/com/scylladb/cdc/cql/driver3/Driver3FromLibraryTranslator.java
@@ -96,7 +96,7 @@ public class Driver3FromLibraryTranslator {
             }
             case DURATION: {
                 CqlDuration libraryDuration = (CqlDuration) libraryObject;
-                return Duration.newInstance(libraryDuration.getMonths(), libraryDuration.getDays(), libraryDuration.getNanoseconds());
+                return Duration.newInstance((int) libraryDuration.getMonths(), (int) libraryDuration.getDays(), libraryDuration.getNanoseconds());
             }
             case DATE: {
                 CqlDate libraryDate = (CqlDate) libraryObject;
